% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/GetServices.R
\name{GetServices}
\alias{GetServices}
\title{GetServices}
\usage{
GetServices()
}
\description{
This function gets the table of web services from the HIS Central catalog
}
\examples{
GetServices()
}
\keyword{waterml}

