% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetSites.R
\name{GetSites}
\alias{GetSites}
\title{GetSites}
\usage{
GetSites(server)
}
\arguments{
\item{server}{The URL of the web service ending with .WSDL,
for example: http://icewater.usu.edu/MudLake/cuahsi_1_0.asmx?WSDL
alternatively this can be the REST URL to get the sites.}
}
\description{
This function gets the table of sites from the WaterML web service
}
\examples{
GetSites("http://icewater.usu.edu/MudLake/cuahsi_1_0.asmx?WSDL")
}
\keyword{waterml}

