% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetValues.R
\name{GetValues}
\alias{GetValues}
\title{GetValues}
\usage{
GetValues(server, siteCode, variableCode, startDate = NULL, endDate = NULL,
  daily = NULL)
}
\arguments{
\item{server}{The URL of the web service ending with .asmx,
for example: http://worldwater.byu.edu/interactive/rushvalley/services/index.php/cuahsi_1_1.asmx?WSDL}

\item{siteCode}{The site code. To get a list of available site codes, see GetSites() function
and use the FullSiteCode field.}

\item{variableCode}{The variable code. To get a list of possible variable codes, see GetVariables()
function and use the FullVariableCode field}

\item{startDate}{(optional) The start date in "yyyy-mm-dd" format}

\item{endDate}{(optional) The end date in "yyyy-mm-dd" format}

\item{daily}{(optional) If you set daily="max", daily="min" or daily="mean", then the
data values are aggreagted to daily time step.}
}
\value{
a data.frame of data values with the following columns:
\itemize{
\item time: The local date/time of the observation. The data type is POSIXct.
\item DataValue: The observed data value
\item UTCOffset: The difference between local time and UTC time in hours
\item CensorCode: The code for censored observations. Possible values are nc (not censored),
            gt (greater than), lt (less than),
            nd (non-detect), pnq (present but not quantified)
\item DateTimeUTC: The UTC time of the observation. The data type is POSIXct.
\item MethodCode: The code of the method or instrument used for the observation
\item SourceCode: The code of the data source
\item QualityControlLevelCode: The code of the quality control level. Possible values are
            -9999 (Unknown), 0 (Raw data), 1 (Quality controlled data),
            2 (Derived products), 3 (Interpreted products), 4 (Knowledge products)
}
}
\description{
This function gets the time series data values from the WaterML web service
}
\examples{
#example 1: Get Values from a known site and variable from RushValley server
v <- GetValues("http://worldwater.byu.edu/app/index.php/rushvalley/services/cuahsi_1_1.asmx?WSDL",
           site="Ru5BMMA", variable="SRS_Nr_NDVI", startDate="2014-11-01", endDate="2014-11-02",
           daily="max")
#example 2: Get values from an external REST URL (in this case the Provo USGS NWIS site id 10163000)
url<-"http://waterservices.usgs.gov/nwis/dv/?format=waterml,1.1&sites=10163000&parameterCd=00060&
startDT=2015-01-01"
v <- GetValues(url)
}
\keyword{waterml}

