% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetVariables.R
\name{GetVariables}
\alias{GetVariables}
\title{GetVariables}
\usage{
GetVariables(server)
}
\arguments{
\item{server}{The URL of the web service ending with ?WSDL,
for example: http://worldwater.byu.edu/app/index.php/rushvalley/services/cuahsi_1_1.asmx?WSDL}
}
\description{
This function gets the table of variables from the WaterML web service
}
\examples{
GetVariables("http://worldwater.byu.edu/app/index.php/rushvalley/services/cuahsi_1_1.asmx?WSDL")
}
\keyword{WaterML}

