% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PackageDocumentation.R
\docType{package}
\name{WaterML-package}
\alias{WaterML-package}
\title{Fetch and analyze data from WaterML or CUAHSI WaterOneFlow web service}
\description{
\tabular{ll}{
Package: \tab WaterML\cr
Type: \tab Package\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
}
}
\details{
Fetch and analyze data from WaterML or CUAHSI WaterOneFlow web service.

Please see \url{https://github.com/jirikadlec2/waterml} for more information.
}
\author{
Jiri Kadlec \email{jirikadlec2@gmail.com}
}
\keyword{CUAHSI,}
\keyword{HydroServer,}
\keyword{WaterML,}
\keyword{services}
\keyword{web}

