% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterOneFlowNamespace.R
\name{WaterOneFlowNamespace}
\alias{WaterOneFlowNamespace}
\title{WaterOneFlowNamespace}
\usage{
WaterOneFlowNamespace(version)
}
\arguments{
\item{version}{The version of the WaterOneFlow XML namespace. Must be either "1.0" or "1.1"}
}
\value{
A list with the namespaces and corresponding prefixes. This namespace
information is important for correct parsing of the WaterML XML document.
}
\description{
A helper function that finds out the WaterOneFlow namespace information
based on the version number 1.0 or 1.1.
}
\examples{
ns <- WaterOneFlowNamespace("1.0")
ns <- WaterOneFlowNamespace("1.1")
}
\keyword{WaterML}
