% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterOneFlowVersion.R
\name{WaterOneFlowVersion}
\alias{WaterOneFlowVersion}
\title{WaterOneFlowVersion}
\usage{
WaterOneFlowVersion(WSDL)
}
\arguments{
\item{WSDL}{The URL of the WSDL, for example http://icewater.usu.edu/MudLake/cuahsi_1_0.asmx?WSDL}
}
\value{
A list with two items: Version (either 1.0 or 1.1), and Namespace
(either http://www.cuahsi.org/his/1.0/ws/ or http://www.cuahsi.org/his/1.1/ws/)
}
\description{
A helper function that finds out the WaterOneFlow service version from
the URL of the wsdl file. By default it checks for "cuahsi_1_0" or "cuahsi_1_1" in
the url. If that is not found, then the function checks the version inside the
WSDL file.
}
\examples{
versionInfo <- WaterOneFlowVersion("http://icewater.usu.edu/MudLake/cuahsi_1_0.asmx?WSDL")
versionInfo$Version
versionInfo$Namespace
}
\keyword{WaterML}
