% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetSites.R
\name{GetSites}
\alias{GetSites}
\title{GetSites}
\usage{
GetSites(server)
}
\arguments{
\item{server}{The URL of the web service ending with .asmx,
for example: http://worldwater.byu.edu/interactive/rushvalley/services/index.php/cuahsi_1_1.asmx}
}
\description{
This function gets the table of sites from the WaterML web service
}
\examples{
GetSites("http://worldwater.byu.edu/interactive/rushvalley/services/index.php/cuahsi_1_1.asmx")
}
\keyword{waterml}

