% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/GetValues.R
\name{GetValues}
\alias{GetValues}
\title{GetValues}
\usage{
GetValues(server, site, variable, startDate, endDate, daily = NULL)
}
\arguments{
\item{server}{The URL of the web service ending with .asmx,
for example: http://worldwater.byu.edu/interactive/rushvalley/services/cuahsi_1_1.asmx}

\item{site}{The site code. To get a list of available site codes, see GetSites() function}

\item{variable}{The variable code. To get a list of possible variable codes, see GetVariables()}

\item{startDate}{The start date in "yyyy-mm-dd" format}

\item{endDate}{The end date in "yyyy-mm-dd" format}

\item{daily.}{Defaults to NULL. If you set daily="max", daily="min" or daily="mean", then the
data values are aggreagted to daily time step.}
}
\description{
This function gets the time series data values from the WaterML web service
}
\examples{
GetValues()
}
\keyword{waterml}

