% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/GetValues.R
\name{GetValues}
\alias{GetValues}
\title{GetValues}
\usage{
GetValues(server, siteCode, variableCode, startDate, endDate, daily = NULL)
}
\arguments{
\item{server}{The URL of the web service ending with .asmx,
for example: http://worldwater.byu.edu/interactive/rushvalley/services/index.php/cuahsi_1_1.asmx}

\item{siteCode}{The site code. To get a list of available site codes, see GetSites() function
and use the FullSiteCode field.}

\item{variableCode}{The variable code. To get a list of possible variable codes, see GetVariables()
function and use the FullVariableCode field}

\item{startDate}{The start date in "yyyy-mm-dd" format}

\item{endDate}{The end date in "yyyy-mm-dd" format}

\item{daily}{Defaults to NULL. If you set daily="max", daily="min" or daily="mean", then the
data values are aggreagted to daily time step.}
}
\description{
This function gets the time series data values from the WaterML web service
}
\examples{
#example 1: Get Values from a known site and variable from RushValley server
GetValues("http://worldwater.byu.edu/interactive/rushvalley/services/index.php/cuahsi_1_1.asmx",
           site="Ru5BMMA", variable="SRS_Nr_NDVI", startDate="2014-11-01", endDate="2014-11-21",
           daily="max")

#example 2: Get values from a random site and random variable
server <- "http://hydrodata.info/chmi-h/cuahsi_1_1.asmx?wsdl"
sites <- GetSites(server)
randomSite <- sites[sample(nrow(sites),1),]
variables <- GetVariables(server)
randomVariable <- variables[sample(nrow(variables),1),]
values <- GetValues(server, randomSite$FullSiteCode,
          randomVariable$FullVariableCode, Sys.Date()-365, Sys.Date())
if(!is.null(values)) {
 plot(values, type="l", main=randomSite$SiteName,
 ylab=paste(randomVariable$VariableName,
 randomVariable$UnitAbbreviation))
}
}
\keyword{waterml}

